+++ packages/paraglide/paraglide-js-adapter-sveltekit/src/test.svelte
@@ -1,0 +1,2 @@
+<!-- svelte-ignore a11y-missing-attribute --><a >ASD</a>
+<!-- svelte-ignore a11y-missing-attribute --><a >ASD</a>